requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.5=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=py38h06a4308_0
attrs=23.1.0=py38h06a4308_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
blosc2=2.0.0=pypi_0
bottleneck=1.3.7=py38ha9d4c09_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.6.2=py38h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
comet-ml=3.43.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
configobj=5.0.8=pypi_0
contourpy=1.1.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
cython=3.0.10=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dgl=2.2.1.th22.cu118=py38_0
dulwich=0.22.1=pypi_0
easytorch=3.8.5=pypi_0
einops=0.8.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
everett=3.1.0=pypi_0
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.53.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py38h5eee18b_0
fsspec=2024.3.1=py38h06a4308_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.31.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.64.1=pypi_0
huggingface-hub=0.23.4=pypi_0
idna=3.7=py38h06a4308_0
imageio=2.34.1=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=py38h06a4308_0
joblib=1.4.0=py38h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lightning=2.2.5=pyhd8ed1ab_0
lightning-utilities=0.11.2=pyhd8ed1ab_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
msgpack=1.0.8=pypi_0
multidict=6.0.4=py38h5eee18b_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pypi_0
opencv-contrib-python-headless=4.10.0.84=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.14=h5eee18b_0
packaging=23.2=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py38h5eee18b_0
pip=24.0=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.10.0=py38h06a4308_0
pooch=1.7.0=py38h06a4308_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=5.27.2=pypi_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pyg=2.5.2=py38_torch_2.2.0_cu118
pyg-lib=0.4.0+pt22cu118=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.0.9=py38h06a4308_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-box=6.1.0=pypi_0
python-dateutil=2.9.0post0=py38h06a4308_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.2.0=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.2.2=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=py38h5eee18b_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.31.0=py38h06a4308_1
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=py38h1128e8f_1
scipy=1.10.1=py38hf6e8229_1
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
sentry-sdk=2.3.1=pypi_0
setuptools=69.5.1=py38h06a4308_0
simplejson=3.19.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py38h06a4308_0
tables=3.8.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.7.10=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
torch-sparse=0.6.18+pt22cu118=pypi_0
torch-summary=1.4.5=pypi_0
torchaudio=2.2.0=py38_cu118
torchdata=0.7.1=pypi_0
torchmetrics=1.4.0.post0=pyhd8ed1ab_0
torchtriton=2.2.0=py38
torchvision=0.17.0=py38_cu118
tornado=6.1=py38h0a891b7_3
tqdm=4.66.4=py38h2f386ee_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.41.2=pypi_0
typing-extensions=4.11.0=py38h06a4308_0
typing_extensions=4.11.0=py38h06a4308_0
urllib3=2.2.1=py38h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py38h06a4308_0
wrapt=1.16.0=pypi_0
wurlitzer=3.1.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.3=py38h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zipp=3.19.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
